requirements.txt
# MSME Shield – Active MVP Dependencies
# Install with: pip install -r requirements.txt

# --- Core API & Validation ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
email-validator==2.1.1

# --- Frontend Dashboard ---
streamlit==1.35.0
pandas==2.2.2        

# --- HTTP Client ---
requests==2.32.3

# --- Security & Reporting (Active MVP Modules) ---
python-nmap==0.7.1
reportlab==4.2.0
zxcvbn==4.4.28
psutil==7.0.0

#---Database---
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# --- Auth ---
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1